Split edit task modal into own file
